LayoutTests/http/tests/media/fairplay/fps-sessionId-expected.txt

PROMISE: requestMediaKeySystemAccess resolved
PROMISE: createMediaKeys resolved
FETCH: server certificate recieved
PROMISE: arrayBuffer resolved
PROMISE: setServerCertificate resolved
EVENT(message)
PROMISE: licenseResponse resolved
PROMISE: session.update() resolved
EXPECTED (session.sessionId != '') OK
END OF TEST
